implement AbstractSelectionUserInput class to replace Selection + SelectionFreeText interface
It make more sense to have a abstract class rather than the interface because Selection and SelectionFreeText variation of UserInput is a is-a relationship. So rather to have duplicate implementation across all the userinput classes, an abstract class it much better.
