+++ src/main/java/com/bwgjoseph/springmvcdynamicuserinput/PersonController.java
@@ -1,5 +1,9 @@
 package com.bwgjoseph.springmvcdynamicuserinput;
 
+import java.util.Optional;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,6 +19,11 @@ public class PersonController {
     public PersonController(PersonRepository personRepository) {
         this.personRepository = personRepository;
         this.personDTOtoPersonConverter = new PersonDTOtoPersonConverter(personRepository);
+    }
+
+    @GetMapping("person/{id}")
+    public Optional<Person> get(@PathVariable String id) {
+        return this.personRepository.findById(id);
     }
 
     @PostMapping("person")
