+++ src/main/java/com/bwgjoseph/springmvcdynamicuserinput/notinuse/Selection.java
@@ -1,0 +1,7 @@
+package com.bwgjoseph.springmvcdynamicuserinput.notinuse;
+
+import com.bwgjoseph.springmvcdynamicuserinput.userinput.UserInput;
+
+public interface Selection<T extends Enum<T>> extends UserInput {
+    T getSelectionValue();
+}
