+++ src/main/java/com/bwgjoseph/springmvcdynamicuserinput/notinuse/SelectionFreeText.java
@@ -1,0 +1,7 @@
+package com.bwgjoseph.springmvcdynamicuserinput.notinuse;
+
+import com.bwgjoseph.springmvcdynamicuserinput.userinput.UserInput;
+
+public interface SelectionFreeText<T extends Enum<T>> extends UserInput {
+    T getSelectionValue();
+}
