--- src/main/java/com/bwgjoseph/springmvcdynamicuserinput/userinput/Selection.java
@@ -1,5 +1,0 @@
-package com.bwgjoseph.springmvcdynamicuserinput.userinput;
-
-public interface Selection<T extends Enum<T>> extends UserInput {
-    T getSelectionValue();
-}
