--- src/main/java/com/bwgjoseph/springmvcdynamicuserinput/userinput/SelectionFreeText.java
@@ -1,5 +1,0 @@
-package com.bwgjoseph.springmvcdynamicuserinput.userinput;
-
-public interface SelectionFreeText<T extends Enum<T>> extends UserInput {
-    T getSelectionValue();
-}
